Moved common files out of website/static/common to website/common. Moved media files out of website/static/media to website/media.
Updated base.html and settings files. Added gitignore for both website/static and website/media.

This better aligns the production, test, and development settings files and will allow us to eventually combine the settings files into a single file again. This has also seemed to have fixed the problem of staticfiles not updating properly if there is a new commit.